+++ nfapi/open-nFAPI/pnf/src/pnf_p7.c
@@ -36,14 +36,14 @@
 #include <SCHED_NR/phy_frame_config_nr.h>
 
 extern int sf_ahead;
-extern void log_mmap_entry(const char *log_name, long value);
+extern void log_mmap_entry(const char *log_name, uint64_t value);
 
-static inline long pack_sfn_slot_value(uint16_t sfn, uint16_t slot, int32_t signed_value)
+static inline uint64_t pack_sfn_slot_value(uint16_t sfn, uint16_t slot, int32_t signed_value)
 {
     uint64_t packed = ((uint64_t)sfn << 48) |
                       ((uint64_t)slot << 32) |
                       ((uint32_t)signed_value);
-    return (long)packed;
+    return packed;
 }
 
 // Used by the RFC3550 jitter calculation (defined later in this file)
